requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
pandas>=1.3.0,<2.0.0
opencv-python>=4.5.0,<5.0.0
Pillow>=9.0.0,<11.0.0
tqdm>=4.62.0,<5.0.0
matplotlib>=3.4.0,<4.0.0
seaborn>=0.11.0,<1.0.0
scikit-learn>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0

# Image augmentation
albumentations>=1.3.0,<2.0.0

# Deep learning (PyTorch)
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
timm>=0.9.0,<1.0.0

# Web UI
streamlit>=1.20.0,<2.0.0

# Logging and experiment tracking
tensorboard>=2.12.0,<3.0.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0

# Utilities
click>=8.0.0,<9.0.0
python-json-logger>=2.0.0,<3.0.0
